+++ seamless/src/serde.rs
@@ -1,2 +1,0 @@
-// These are exposed for our macros to use:
-pub use serde::{ Serialize, Deserialize };